src/requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
pluggy==1.4.0
PyMySQL==1.1.0
pytest==8.0.2
pytest-flask==1.3.0
python-dotenv==1.0.1
SQLAlchemy==2.0.28
tomli==2.0.1
typing_extensions==4.10.0
Werkzeug==3.0.1
